+++ src/AnimeFeedManager.Functions/Models/SeasonInfo.cs
@@ -3,6 +3,6 @@
     public class SeasonInfo
     {
         public string? Season { get; set; }
-        public ushort Year { get; set; }
+        public int Year { get; set; } // Azure tables only works with Int and Int64
     }
 }
