+++ src/AnimeFeedManager.Storage/Domain/AnimeInfoStorage.cs
@@ -8,7 +8,7 @@ namespace AnimeFeedManager.Storage.Domain
         public string? Title { get; set; }
         public string? Synopsis { get; set; }
         public string? FeedTitle { get; set; }
-        public ushort Year { get; set; }
+        public int Year { get; set; } // Azure tables only works with Int and Int64
         public string? Season { get; set; }
         public DateTime? Date { get; set; }
     }
